+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob/master/CHANGES)._
 
-#### 1.9.11-SNAPSHOT
+#### 1.9.11-b1
 
 - **[UPDATE]** Updated to LibGDX 1.9.11.
 - **[UPDATE]** Updated to Kotlin Coroutines 1.3.6.